refactor(scoring): Add scoring helper. (#373)
